First cut at replacing INPUT with improved ASK dialect
This adds the feature that ASK can be passed a TEXT! prompt, a filter
DATATYPE! which the typed-in answer must TO-convert to, or a BLOCK!
containing a prompt and a datatype together.

Because ASK of a DATATYPE! does not prompt, the sequence ASK TEXT!
replaces the functionality of INPUT.  While the plan is to remove
INPUT ultimately, for now this just makes it a synonym in order to
test the feature and give people some time to test their code.  The
functionality previously given to INPUT is now hookable as READ-STDIN.

Improvements to this simple first-cut at the dialect are being
discussed:

https://forum.rebol.info/t/1124