Standardize Rebol comments: no `;;` or `;--`
This goes ahead and changes the file to use a standard comment format
of single semicolon, two spaced off:

    some [code]  ; a comment

This takes the place of "noisier" approaches:

    some [code] ;; a comment

    some [code] ;-- a comment

The whitespace provides a more natural division for the eye, without
the clutter.  It matches a parallel spacing technique now used for C:

     some(code);  // a comment

This change unearthed some accidental double-semicolon cases in
terminating lines in the C, and cleans those up.

Also, some commented out code was either removed or changed to use
COMMENT [].

Also adds to the tests the ability to allow a URL! in a test group
as an annotation, similar to how ISSUE! is allowed.  This seems needed
because not all "see this" links correspond to issue numbers that map
to the issue database (for instance, Ren-C has its own issue tracker)
but also because the "see this" link might be something else, like
a commit.