Don't use LOAD for most bootstrap files (mezz/host-startup)
Historically it has been very difficult to leverage changes in the
scanner or anything that would affect LOAD for code in the mezzanine.
This is because it relied on being LOAD-ed and MOLD/FLAT to strip
out the comments and whitespace from the file.

A few files used the state of being LOADed for more, e.g. structural
inspection to produce C constants corresponding to definitions in the
files.  But the vast majority of code was simply using it to remove
the comments and whitespace.

This commit implements a long-desired facility via a STRIPLOAD function
that takes out comments and whitespace, but does not use LOAD.  It's
not entirely trivial--since it involves (for instance) knowing whether
you are in a multi-line comment or not--accounting for things like
nested braces.  However, this can be done with relatively easy to
understand PARSE rules.

In order to get the bootstrap sorted out, this brings back the IF
construct for PARSE.  It will be replaced by `:(...)` in the next
bootstrap executable pick.  But because it is back, it's now possible
to get rid of the `((...))` behavior in PARSE for the same feature.