Use SYM-WORD! for fail, and in %prot-http.r
During the update of TLS, the coming existince of @word was imagined...
and so EMAIL! was used to try and diversify the types (but as `word@`).
No longer dependent on the properties of an old Rebol's LOAD for
bootstrap, this goes ahead and uses the real thing.

Also, one of the first envisionments of the use of @ was to say where
a failure occurred.  With the bootstrap change, that can now be put
into effect and say `fail @location` in the mezzanine.