main.c only uses libRebol, clarify/organize r3.exe build
Originally the %host-main.c file did not depend on %sys-core.h, but
only included the RL_API files.  In order to open up the development
of the console to have non-trivial features, it was extended to have
access to the full Rebol API (while trying to be forward-looking and
using features minimally).

The code is now refactored, including the factoring of the console
behavior into an extension (that is reused in the JavaScript build).
With a full API available via libRebol, the necessary usermode
routines for binding and loading now can be accessed.  So this goes
ahead and uses them vs. reaching into %sys-core.h, removing the only
remaining dependency on anything other than %rebol.h.

Several known open issues pertaining to module usage and binding are
involved.  This attempts to document and structure the code to make
those more comprehensible.  It also moves the files into a clearly
marked "main" directory--where the prep script for embedded data lives
there too, making it clear to see how the parts fit together.