Fix #prefer-O2-optimization for MSVC build
The MSVC Visual Studio solution generator handles optimizations
differently, so it didn't pick up on the "per-file-override" hack that
is used to make %c-eval.c optimize for speed even in a size-optimized
build (which is important for the WASM build).

This works around the issue well enough to get Visual Studio solutions
generating again.