Move CHECKSUM to Crypt Extension
Zlib provides a CRC32 and ADLER32 checksum.  R3-Alpha also built in MD5
and SHA1 (having apparently once supported MD4 as well, but that was
not in the released code).

Modern cryptography has generally passed the MD5 and SHA1 algorithms,
making them not great choices to force everyone using the interpreter
to carry along.  For the moment, this goes ahead and moves the
checksum function itself into the Crypt extension...though exposure
of the ADLER32 and CRC32 might make sense from core so long as they
are built in.

A future method would hopefully allow the checksums to be each their
own extension you could select to be in the build or not (or to get
dynamically.)