Some improvement to UTF-8 Everywhere AS conversions
This moves along the UTF-8 Everywhere conversions of BINARY! to an
ANY-STRING! by mapping to a correct index (and making sure that the
binary is on a codepoint position).

Adds a few tests (though not nearly as many as needed!) for the AS
behaviors, including the somewhat unusual one that an AS conversion of
a BINARY! to UTF-8--if it succeeds--will constrain the BINARY! so that
modifications can only produce more UTF-8.