Expose CRC-32 and ADLER-32 from zlib in the core
Hashes like MD4, MD5, or SHA1 require tables and algorithms and take
up space.  Yet these fall out of fashion, and may not be required for
all builds.

All cryptographic hashes were thus moved to the Crypt extension (and
should probably be more modular than that, so that they can be
included in a build or dynamically on an as-needed basis).  However,
the CRC-32 hash was used in the unzip script which is part of the
core build, not bound to the crypt extension.  It also was called
by Mezzanine save code.

The CRC-32 hash is used in Gzip, which is the format chosen to compress
embedded code in Rebol.  Hence it is a sunk cost in the executable.
This exposes it via a CHECKSUM-CORE routine...and also exposes the
ADLER-32 hash from zlib.

Because of its narrow usages, this goes ahead and experiments with
avoiding whether to return a signed or unsigned integer by returning
bytes.  The bytes are in little endian format, which the relevant
usages of CRC-32 expect.