Finalize QUOTE => evaluative quote level adder
Historical Rebol called the operation that would literally give you
a parameter QUOTE:

    rebol2>> quote (a)
    == (a)

The apostrophe before certain types which suppressed their evaluation
was called "lit"

    rebol2>> type of first ['a]
    == lit-word!

Ren-C has shuffled the terms around a bit due to "generalized quoting",
so that 'a is actually a *quoted* WORD!, not a LIT-WORD!.  A "literal
word" is the word itself.  e.g.:

    ren-c>> literal x
    == x

    ren-c>> lit (a)
    == (a)

Then, quoting adds a level to things

    ren-c>> quote first [a]
    == 'a

    ren-c>> quote quote lit (a)
    == ''(a)

Since this is a significant change, it was phased in.  Since QUOTE
takes the place of a previous helper called UNEVAL, the UNEVAL name
was kept for some time while QUOTE gave an error.  However, as this
is now finalized, the operation takes its final name.