Move SECURE dialect and support to an extension
R3-Alpha intended to expand upon Rebol2's privileges model in a generic
way, with a SECURE dialect that offered more than restricted access
to files.  This wasn't finished, and security is not something the
language model has built from foundationally with its evaluation model.
Yet it was being carried around in the executable as code both in the
core and in the mezzanine.

The issue may be best addressed outside the language level (and at the
OS or virtualization level, broadly in a cross-cutting way that would
apply equally to other languages or tools).  However, it is also true
that a little bit of foresight on such features could be useful--
especially in constrained environments where the systemic features
are not implemented.

This preserves the code, but moves it into an extension--keeping it
in the desktop builds, but not the web build.