Shutdown devices in API layer, not Shutdown_Core()
The goal of Shutdown_Core() is to free all memory used by the
interpreter (and it checks that it does--even if you are not using
something like ASAN or Valgrind).  However, it shouldn't be necessary
to call it if one is not going to call Startup_Core() again (and not
doing a debug build where you're checking for memory leaks).

In an attempt to remove all "internal" API calls from the %main.c,
however, the call to OS_Quit_Devices() was moved to Shutdown_Core().
This wasn't the right place for it...as release builds then don't run
that code.  One side effect of this was that the terminal would stick
in non-echo mode that it had been put into to achieve line editing on
the POSIX builds.

This moves the device quitting code into the API-layer of shutdown
that happens regardless of whether it's a debug build or not.