Give types of refinement arguments in HELP
With the change to where "refinements are the name of the refinement
argument themselves", this means that HELP needs to show the type
information for that argument.  As it happens, the types would have
shown if they hadn't been suppressed.

This removes the suppression, so the types show on the same line as
the refinement (as it would for a non-refinement argument).