When `<script>` used, run code in reb.Promise()
While it is technically possible to run Rebol code on the "gui thread"
in JavaScript, routines that interoperate with the DOM or use async
functions with `await` must run the code on a worker thread, so that
it can be suspended while the GUI thread runs.

The fledgling ability to run Rebol code via a `<script>` tag in HTML
source was using `reb.Elide()` to do it on the GUI thread.  This
changes that over to use a `reb.Promise()`.