Make PATH! composition tolerate WORD!, plus bugfix
COMPOSE of a PATH! was broken.  This fixes one major issue with it,
and also relaxes COMPOSE to be willing to take ANY-WORD! and pass it
through (which Rebol2 did as well).  This way you can do a SET/HARD
of a COMPOSE'd path-or-word, thus conveniently getting past the rule
that SET and GET won't evaluate GROUP!s in paths implicitly.