Fix date normalization when no time component present
There is a distinct state in Ren-C for not having a time from having
a time component is zero.  This uses a special out of band value.
However, that value was not accounted for when dates were being
normalized...which meant a very strange answer was given back when
date math would otherwise produce a nonsense date, e.g.

    d: 29-Feb-2020
    d/year: d/year - 1

This doesn't do time normalization when the component is indicated
as not being present on a date.