Eliminate extra TO BINARY! on LOAD of TEXT! source
Generically speaking, the scanning mechanics work on bytes that have
not been pre-confirmed as UTF-8 (or which can even be embedded in
a larger BINARY! which only has a portion of content that is valid
UTF-8...the portion being scanned).

In R3-Alpha this meant scanning related code could not run on a string
directly, because it was using UCS-2 instead of UTF-8.  But in Ren-C
there is now UTF-8 everywhere.

This means that if you LOAD a TEXT!, it does not have to have a copy
made as a BINARY!.  It can be aliased with AS.  This saves time and
space in the loading process if it is going by way of text.

This makes that change, and also includes a tweak to HTTP to not make
an extra copy of textual payloads as binary either.  Also invalidates
bookmarks when doing DELINE (as character offsets may change).