Faster generic JS-EVAL* low-level evaluation native
The first hack for implementing a string eval in the JavaScript
extension was to create a one-off JS-NATIVE whose body was the string
(but took no arguments), and call it once.  That was terribly
inefficient, but took care of the concerns of moving the code to the
appropriate thread.

That did not allow for global evaluation, however...since proxying code
to a main thread always involved running inside a function.  And it
seemed that if you called eval() from in a function you could not add
global variables as `var x = ...` (you'd have to say `window.x = ...`)
But since code existed for JS-DO'ing URLs, that code for using a
`<script>` tag was adapted to add a script tag and await on notice of
that script tag finishing.  This was also terribly inefficient.

This commit solves both of those issues by adding a significantly
faster low-level JS-EVAL* primitive.  It takes advantage of a trick
that *indirect* calls to `eval` will be global, e.g.

     eval("var x = 10")  ; local
     (1,eval)("var x = 10")  ; global

For more details on that:

http://perfectionkills.com/global-eval-what-are-the-options/

So JS-EVAL* defaults to global evaluation, but can also be asked to
do /LOCAL evaluation.  With /VALUE it can be asked to convert the
result to a Rebol value (currently only covering NULL, VOID! (e.g
undefined) INTEGER! and TEXT!...with all other results returning
void at this time, though it may be changed to raise an error).

Hence it is now the workhorse behind the higher level JS-DO when it
takes a TEXT! (which returns no result, and also works with URL!)...
as well as JS-EVAL, which is presumed to evaluate locally and return
a result.  Both of the higher level constructs can take a BLOCK! to
splice in Rebol values using special forms.