Fix DELTA-TIME
The DELTA-TIME function used a trick to be more accurate, by getting
rid of variable assignments and using a two-phase process (first a
REDUCE and then a DO).  Yet it actually messed up by running the
timings in the second phase, without bracketing the code (!).

This fixes the issue and enhances the trick slightly.