Add libRebol symbols with tcc_add_symbol()
Although Linux automatically exposes symbols from the host Rebol
executable to code that is built dynamically and run by TCC in memory
(even ones it arguably shouldn't...), Windows doesn't do this.  So the
tcc_add_symbol() function must be called to expose entry points from
the Rebol executable to code from a user native.

This adds the symbols for libRebol.  It may be that the table of entry
points that is built for libRebol could be used to make the symbol
list instead of a different prep step during the build...but that would
mean storing the string names of the entry points as well as their
function pointers.