Don't use package manager tcc on trusty Linux (2013)
Although it should theoretically be possible to use the `sudo apt-get`
variant of tcc on 64-bit linux when building for 64-bit linux, in
practice there's something wrong with the version in the trusty
package manager.  It gives incompatible results between GCC's va_list
(which is what the interpreter runtime expects) and what the TCC
helpers in libtcc1.a are producing.

The incompatibility doesn't seem to exist in modern TCC, so the easiest
thing to do is to just build it from source (which already needed to
be done for cross compilation anyway).