Support `some-file/bar/` path construction
The concept of producing new FILE! paths by PATH!-ing a FILE! is
somewhat sketchy.  But if it exists, it should be able to support
FILE! with a terminal slash...as that is considered to be Rebol and
Red's standard representation.

https://github.com/rebol/rebol-issues/issues/2378

This can't be done with PATH! as conceived in R3-Alpha or Red.  But
Ren-C treats BLANK! in path as an invisible slot (as part of the
unification of PATH! and REFINEMENT!).  That affords terminal slashes
as well.

Also fixes a scanner issue relating to slash-terminal paths which
are at the end of GROUP!s or BLOCK!s.