Fix unzip.reb handling of end of input block
Unzip was running afoul of the situation in MATCH where a falsey
input which matches (e.g. blank) returns VOID!, in order to avoid
a deceptive negative result that would suggest no match.

This rewrites the condition to instead of looking for cases that
mean the subsequent item should *not* be treated as content of the
filename, to instead match the cases that *are the content*.

Also fixes the /VERBOSE option, as PRINT no longer accepts FILE!.

Includes adjustment to the previous commit to account for line endings
or end of input when PATH! ends in a slash.