Make tcc-encap.zip file for using TCC without dependencies
This has the build pack up a small zip of support files that are needed
to use with a Rebol that has been built with the TCC extension in
order to write user natives.  Ultimately these files should be encapped
into the built executable itself, and then automatically unzipped into
a temporary directory in order to use for TCC building.  But this
is a first step.

Also allows you to specify the LIBREBOL_INCLUDE_DIR and CONFIG_TCCDIR
via compile options.  That's because setting environment variables
isn't always as easy as editing a source file, so this makes it
possible to change %fib.r to point at the directories for %rebol.h and
for the TCC runtime.