Make it possible to specify one directory for TCC Windows
Having achieved the first successful build of %fib.r on Windows (that
I have ever seen), this now tries to take that process and make it
work from a single directory's worth of unpacked TCC files.  For the
moment that .zip is a separate download on Travis, but it is moving
closer to being something that could be made turnkey.