Use indirect calls on rebEnterApi_Internal
When an EXE loads an extension DLL that uses libRebol, that DLL speaks
to the API through an indirection of a table.  This is because you
cannot generically link DLLs to call functions in an EXE.

There are thus two forms of the API calls generated: direct calls
(if you are building an EXE and calling Rebol as a static or shared
library), and indirect calls (if you are building a DLL and calling
back into the Rebol EXE that hosts the DLL).

A bug in the indirect formulation was trying to make a direct call to
the generic "entry checking" function for the API.  That function is
intended to help avoid boilerplate calls in %a-lib.c, but perhaps
that could be done with some other macro more efficiently.  Either
way, this fixes up the issue.

Also explicitly mentions REBOL_IMPLICIT_END in the inclusion of
"rebol.h" (even though it is the default for C99)