EVAL => REEVAL, EVALUATE takes @var, @var/path, or @(var expr)
This is an update of the EVALUATE native to support setting variables
in a more convenient notation than /SET and putting the variable after:

    ; old way
    evaluate/set block 'var

    ; new way
    evaluate @var block

It uses skippable parameters, so if you do not provide an @... then
it will just evaluate and return the new position in the block without
setting any variable.

You can also use a PATH! to specify the variable (which DO/NEXT could
not do).  Also, you can use @(...) to specify an expression that can
evaluate to a WORD!, PATH!, or BLANK!.  If it is BLANK! then no
variable will be set.

It also renames the old EVAL to REEVAL.  This is because it's confusing
to have EVAL be anything other than a short way of typing EVALUATE.
Ultimately EVAL=>EVALUATE and REEVAL=>REEVALUATE, but for now EVAL
is an error to catch old uses.

https://forum.rebol.info/t/eval-evaluate-and-reeval-reevaluate/1173