Re-enable PROBE() of PARAMLISTs
PROBE() on an action's parameter list was asserting due to the use
of > REB_MAX values for the VAL_TYPE() of params.  This changes it to
use the MIRROR_BYTE() which holds the underlying TYPESET! type for
the cell itself.

(The technique is likely to change, but this just gets past the crash
in the debugging for now)