Add POSIX C99 compilation command to TCC native
The TCC extension was initially written to just compile user natives
into memory and execute them.  However, TCC is also able to build and
link executable files on disk.

This adds a `c99` usermode function to the TCC extension to implement
the POSIX protocol for c99 command lines:

http://pubs.opengroup.org/onlinepubs/9699919799/utilities/c99.html

In order to do this, it extends the low level COMPILE* native to
support input lists of files (e.g. multiple object files) as well as
an OUTPUT-FILE and OUTPUT-TYPE.