Eliminate printf-like "Emit()" function
Some of the molding methods depended on a printf-like function called
"Emit()".  This had to interpret a format string, and proceeded
character by character and process arguments in a va_list.  The uses
were limited, with some format specifiers being used by only one
caller...making the true behavior less comprehensible than calling the
functions directly.

Additionally, the types of the variadic arguments weren't checked at
compile-time the way a normal function argument would be.  This made
bugs harder to find, and changes more difficult to make.  With the
advantages limited compared to the downsides, it's better to get rid of
it, and revisit the question of more efficient factoring.

One change in this commit is to switch LOGIC! from molding as a WORD!
to showing as #[true] and #[false].  It still FORMs as the WORD!s,
however.