REBCNT => REBLEN
The "Rebol Count" has an unfortunate pronunciation problem; and what it
actually represens is an unsigned length in units (e.g. values in
an array) as opposed to a byte size (which is covered by REBSIZ).

Calling it REBLEN is clearer and matches the terminology of the system
(e.g. VAL_LEN_AT() gives back a REBLEN).

Includes a few changes related to differentiating between sizes and
lengths, including an improvement to the error reporting in the
generation of hash primes.