Fix BACK_CHR() out-of-bounds problem (found with Valgrind)
A problem that Valgrind picked up on that Address Sanitizer did not is
that when enumerating backwards over a string, the final step would
scan for the next codepoint position before the string start.  This
would find a position that would not wind up being used in the next
iteration, however it's still an out-of-bounds memory access with
undefined behavior.

The easiest way to address this is to change BACK_CHR() so that it
actually ignores the character at the current position, and reads the
character before it.  This means you can start from the tail of a
string (on the NUL terminator) and the first character you read will
be the one before the NUL.  This way, the last enumeration will give
you the first character but have your position as the start of the
string data--without scanning before that.

Eliminates STR_LAST(), because the only usages pertained to the buggy
protocol for BACK_CHR() that was stepping out of bounds.