Fix stale pointer in parameter enumeration of specialization
The approach used by enumeration was to synthesize a temporary
parameter for partially specialized refinements that was not marked as
being a refinement.  This worked all right for clients that wanted to
enumerate the arguments and not store that temporary, but the code
for finding the First_Unspecialized_Param() would return that pointer
to the caller...which would not longer be valid.

This changes the enumerator to take a flag as to whether the refinement
status should be ignored on the parameter.  A better solution would
make the specialized paramlist not carry the refinement flag at all,
but that would take some rethinking to still be able to honor the
specialization override in the proper order.