Add library directories in COMPILE step for TCC extension
The TCC command line executable automatically adds library directories
which are not part of the implicit behavior of libtcc.  This hacks in
those directories to the COMPILE front end for the extension.

Also improves error reporting in POSIX command line parsing and adds
support for ignoring a couple of command line switches used in builds.
(Doing so may not be a good idea long term, but is helpful for current
testing scenarios.)