Fix higher quoting levels for unbindable types
Using more than three quote levels on a value can't fit within a value
cell and triggers a separate allocation.  Once a cell takes up actually
two cells, this creates complications for sharing the cell data--as
Rebol historically assumes it can directly manipulate the bindings of
cells and that won't affect another cell that had been copied from it.

The trickery to work around this haad been tested for higher quote
level objects like WORD! or BLOCK! that carry bindings, but not for
unbindable types like INTEGER!.  This repairs two mistakes in that
process.

(Associated test is coming in a separate commit.)