Make macro for UTF-8 continuation bytes
The test for whether something is a UTF-8 continuation byte (assuming
it is in the context of a known valid UTF-8 string) is whether the
byte bitwise and'ed with 0xC0 is 0x80.  There was a misremembered
instance of this test as (>= 0x80), suggesting the various instances
should be macroized to better capture the intent correctly.

The macro also mentions in its name that one shouldn't assume a raw
binary data that hasn't been checked as UTF-8 can be assumed an
extended codepoint--which helps clarify the callsites.