Fix string limit handling in VAL_SIZE_LIMIT_AT
The VAL_SIZE_LIMIT_AT routine returns a length in bytes for the UTF-8
encoding at a certain codepoint index in a string.  It also allowed
for passing in a limit of how may codepoints the length would be for.
If this limit was -1 it would assume no limit.

There was a bug where the limit was not being clipped to the length of
the available codepoints.  This was causing any limit to be passed back
as a valid length, and the return of garbage byte sizes.