Modal arguments firt draft implementation (non-variadic)
This is a basic attempt at implementing modal arguments.  These are
evaluative parameters in which the use of a *literal* @word, @pa/th,
@[bl o ck], and @(gr o up) at a callsite will trigger that element to
be evaluated as if it did not have the @, but automatically flip on the
following refinement.

In order to literally pass the @-arguments, it is possible to do with
quoting...since the arguments are evaluative.  The quote will be removed
and it will be like pasing an @-argument normally.

This property allows distinction from a non-literal argument that is
an @-value, because those do not trigger use of the refinement.

It is rough because it does not pre-check functions to be well formed
(e.g. having a refinement after a modal parameter), instead causing an
error when the function runs.  It also does not include support for
variadic parameters.

The motivating scenario for this feature is to try and solve the chronic
problem of treating a BLOCK! parameter as multiple items by default,
by offering a lightweight single-character way of requesting that
handling.  However, it has potential application to other problems
where there is benefit to marking parameters for needing a special mode
of handling.