Test for unbound words in HELP, misc beneficial tweaks
This commit contains a number of miscellaneous changes, ranging from
typo correction to adding a return annotation to IMPORT, or improving
the readablity of the "big-o" derived binding example.  These changes
were incidental portions of work on other experimental branches.

The commit title mentions the only real functional difference, which
is to test for unbound words in HELP.  The fact that merely typing in
a WORD! would bind it in the user context made this a rare case, but
it could happen, e.g. by explicitly unbinding a word programmatically.

     help (unbind 'foo)