Make reads SYSTEM/SCRIPT/ARGS vs. SYSTEM/OPTIONS/ARGS
The Rebol make+prep process was reading the command line out of
SYSTEM/OPTIONS/ARGS, which apparently is empty if you use DO/ARGS to
invoke the build process from the interpreter itself.

This changes to use the more generic SYSTEM/SCRIPT/ARGS, which will
contain either the DO/ARGS argument or the components of the command
line itself.

It also removes a superfluous PROBE statement, gets rid of the
"SHIMMING OLDER R3" notification that was coming up several times
during the build, and removes a little bit more verbiage from the
build output.