Deeper check for valid UTF-8 in debug build
While the UTF-8 loading code for R3-Alpha gives an error vs. crashing
on certain invalid UTF-8 patterns, it does not *fully* validate UTF-8.
This means some inputs that are actually invalid UTF-8 will just be
loaded as a string of "some other characters":

https://github.com/rebol/rebol-issues/issues/638

The reason for not doing greater validation was based on the belief that
any of the string states you could get to from invalid UTF-8 data would
also be producible from valid UTF-8 data.  Hence, paying a cost (said
to be about 50% overhead) wasn't worth it.

This cost doesn't seem to make sense, as the checks are only on non
ASCII sequences...which are uncommon in typical Rebol code.  The checks
seem quite worth it, and not overly expensive.

So this commit simply reinstates the checks.