Migrate code from unfinished serial extension branch
In the desire to minimize what pieces of the system were necessary to
execute, many "PORT! Actors" and other features were moved into
extensions that could be built a-la-carte.

However, there was an unfinished effort to migrate the serial port out
to an extension that predated the "mass migration".  That older effort
wasn't simply focused on mechanical extraction, but also attempted to
make some changes to how the new API would be used...with the hope
that %sys-core.h would not be necessary, and only the %rebol.h would
be needed.

That more thoughtful extraction attempt ran into walls with the event
pump and device model, which couldn't quite be "wished away" (as they
could in the case of something like the Clipboard).  So it sat on
the side on a branch (multiple branches, which weren't clear what
state they were in, since it had been so long!)

This brings in the API code from what seems to be the latest version of
those branches, so they may be deleted.  It's not clear what the future
of the serial extension is if it does not have a maintainer, but it
could presumably be returned to a working state again if someone
were motivated to do it!