Make tests CHANGE-DIR to test directory when running
This tweaks the behavior of the code that runs tests to switch to the
directory of the test when running it.  As a result, it's possible to
move test directories into locations outside the %tests/ subdir
(e.g. to put tests of extensions in the same directory as the
extension).