Initial transition from EMTERPRETER to ASYNCIFY
Ren-C wants to give the "illusion" of synchronous modification to the
DOM without yielding control of the Rebol stack.  This is possible to
do by running Rebol on a "Web Worker" thread, and then use various
shared memory atomic operations to pause and allow the GUI thread
(which is left free) to handle requests and awake the worker when that
request is fulfilled.

The problem is that web browsers do not necessarily have these features
available yet, or if available do not enable them by default.  To make
sure the web REPL and other Ren-C JS code can run on the widest set
of browsers without needing special configuration, workarounds are
needed.  The Emscripten answer was previously something called the
"Emterpreter" which provided a kind of coroutine simulation through
bytecode, but was substantially slower than threading.

The next version of Emscripten coming down the pipe will be dropping
the Emterpreter for WebAssembly-based builds--keeping it as a last
resort for old browsers that can only run asm.js code.  A new method
known as Asyncify is used, which is significantly faster and has other
advantages.

This is a first cut of just switching to Asyncify.  It's a bit bleeding
edge at the moment--and the Emscritpen SDK for it actually isn't
able to build the Pthreads variant.  So the Travis builds are split...
with Pthreads still using the current SDK, and the Asyncify done on
the latest-upstream.  The idea is to be a bit ahead of the curve and
prepared for the transition.

Full effects of the change are not understood, and there are lingering
comments regarding emscripten which need to be addressed.  However,
getting it deployed in working form has the immediate advantage of
giving a speedup of a factor of about 7x or so vs. the Emterpreter.