Sync up ccall()/cwrap() synchronous call hack for Asyncify
When switching to Asyncify, it wasn't initially known if there was
going to be a diminished need for the workaround that subverted the
assertion about calling back into functions during a yield().  But
the conservative assertions for Asyncify are mirroring the ones
for the emterpreter.

This updates the copied code for the hack to line up with the upstream
version of what it was originally copied from.  It also links to an
issue which has been raised with Emscripten regarding the problem:

https://github.com/emscripten-core/emscripten/issues/9412

Also updates comments on rebIdle_internal() to explain why it is
there vs. just exporting the RunPromise() API directly (which we
could, and could have done before as well.