Improve logic for asynchronous writes
It is desirable to be able to say:

    data: {abc}
    write port data
    reverse data
    write port data

This means that asynchronous writes must copy the data they are given.
In addition to making that copy, it makes those copies API handles
that live with their indidual requests (and are GC managed by those
requests).

In addition, this improves the logic of the WRITE/PART feature, since
it leverages COPY/PART.