Asynchronous errors on PORT!, add READ/PART for TCP
The asynchronous model of ports meant you could not use TRAP on a
READ or WRITE:

     trap [
         write async port data
     ] then [
         ; the write happens much later, so this is ineffective
     ]

This is a very early step toward providing asynchronous error handling,
aiming to make rebol-httpd more robust w.r.t. arbitrary client
disconnections.  While legacy PORT! and device code is not a priority,
the rebol-httpd is an important component of the rebol-server, so this
particular patch makes sense to pay attention to.