Pare down compression entry points in libRebol
The desire to be able to use Rebol's built-in compression (e.g. to
implement PNG encoding and decoding without having a copy of zlib)
led to adding a number of entry points into libRebol to expose that
functionality as C functions.  However, the new direction is to make
the bridging between Rebol and C more seamless--so it should be
possible to efficiently share that functionality by means of
calling natives like "gunzip" via the API and processing the resulting
BINARY! or HANDLE!.

The specifics aren't completely worked out of how bytes might be
exchanged without copying them.  So this leaves behind the two APIs
used by zlib, but switches the approach used in loading the startup
code and the extensions to call into Rebol (the %b-init.c needs to
decompress base/sys/mezz before any code can run, so it now calls the
internal API directly.)