Add some usages of blockscope
It is clarifying in some code that reuses variables to put variables
inside the scope of their applicability, and re-declare them if need
be with new values.  This helps with source understanding and catching
various bugs.

A common pattern was in the Add/Remove binder index cases to reuse
variables where it wasn't clear they were not carrying over meaningful
values from one place to the next.  This adds `blockscope { }` to these
locations (where `blockscope` is actually defined to nothing, but
helps explain why the block is there, and looks a little clearer)