Fold together Copy and Relative Binding into single pass
Ever since R3-Alpha, when a function body was relativized this was
done in two passes: one ran through the generic copying code, and
the next would connect any args or locals to their bindings by
means of the function's identity in those cells.

This goes ahead and implements that as a single pass.  For the moment,
it involves some code duplication of the "Clonify" routine.  However,
the borrowed portions are left intact to make it easier to see how
there might be intelligent code sharing between these routines in
the future.