Eliminate D_ARGC usages
R3-Alpha used integers to access parameters and refinements inside
native routines (including ACTION!s).  While efficient, this impeded
maintenance and readability.

Ren-C kept the idea of accessing via integers, but produced macros
in the prep step so that those integers could be done by parameter
name.  Each native would then include the appropriate macro.  This
made the code more robust with respect to changes, and significantly
more readable.

The situation with ACTION!s was a bit of a problem, because a single
routine had to handle many different function interfaces (by way of
a switch() statement).  The only thing they truly had in common was
the generic first argument of the datatype being dispatched.  However,
other commonalities were exploited by way of using the raw integer
indices.

The most common lingering idea was that binary operations would want
to name an `arg` argument outside the switch.  Increased rigor in
Ren-C would complain if such an argument were accessed for unary
operations, so this noisy assert led to a necessity to know how
many arguments a routine had, leading to instances of:

    REBVAL *arg = D_ARGC > 1 ? D_ARG(2) : NULL;

As it turned out, this was a fairly useless check...as the addition of
RETURN arguments to most actions was meaning the named `arg` argument
was being assigned pretty much all the time.

This commit goes ahead and starts moving the access of `arg` into
basically each branch that needs it, in an unconditional fashion.  It
is once again stopgap, as the entire model for Rebol "generic
functions" needs some deep thought.

Includes some light reformatting in the affected files, so long as
they were being changed.